backend/pyproject.toml
[project]
name = "napling-choice-awards-backend"
version = "0.1.0"
description = "Backend API for Napling Choice Awards"
requires-python = ">=3.12"
dependencies = [
    "flask==3.0.3",
    "flask-cors==4.0.0",
    "flask-limiter==3.5.0",
    "flask-caching==2.1.0",
    "pymongo==4.3.3",
    "python-dotenv==1.0.0",
    "gunicorn==20.1.0",
    "werkzeug==3.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = []

[tool.hatch.build.targets.wheel]
packages = ["."]
